Major enhancements and security updates
Improved api security using tokens and domains and enhanced performance along-with making the data access more secure.

+++ mailer/controllers/templates_controller.py
@@ -2,11 +2,37 @@ from flask import jsonify, make_response, request, g
 from mailer import db
 import datetime
 from mailer.models.templates import Templates
+from mailer.models.app import App
+from mailer.models.user import User
+from mailer.models.mail_log import MailLog
+from mailer.util.encryptor import Encryptor
 from mailer.util.validator import Validator
-from mailer.util.user_token import UserToken
 from mailer.util.html_parser import HTMLParser
 
 class TemplatesController:
+    
+    @staticmethod
+    def validate_template(template_id, user_id):
+        
+        user_data = User.query.filter_by(id=user_id).first()
+        if not user_data:
+            return None
+
+        app_data = App.query.filter_by(user_id=user_id).all()
+        if not app_data or app_data == []:
+            return None
+        
+        apps = []
+        
+        for app in app_data:
+            apps.append(app.app_id)
+        
+        for app in apps:
+            template = Templates.query.filter_by(app_id=app, template_id=template_id).first()
+            if template:
+                return template
+            
+        return None
     
     @staticmethod
     def get_template(template_id):
@@ -15,15 +41,15 @@ class TemplatesController:
             
             user_id = g.user
             
-            template = Templates.query.filter_by(template_id=template_id).first()
+            template = TemplatesController.validate_template(template_id, user_id)
             
             if not template:
                 response = {
-                    "code": 400,
+                    "code": 404,
                     "status": "failure",
-                    "message": "Template does not Exists!"
+                    "message": "Template not Found!"
                 }
-                return jsonify(response), 400
+                return jsonify(response), 404
             
             response = {
                 "code": 200,
@@ -49,6 +75,16 @@ class TemplatesController:
         try:
             
             user_id = g.user
+            
+            app_data = App.query.filter_by(user_id=user_id, app_id=app_id).first()
+            
+            if not app_data:
+                response = {
+                    "code": 400,
+                    "status": "failure",
+                    "message": "Invalid App!"
+                }
+                return jsonify(response), 400
             
             templates_data = Templates.query.filter_by(app_id=app_id).all()
             
@@ -82,21 +118,77 @@ class TemplatesController:
         
         try:
             
-            user_id = g.user
+            req_params = ['app_id', 'name', 'subject', 'sender_name', 'sender_email', 'body', 'is_html']
             
-            if Validator.validate_name(data['name']) == False or Validator.validate_name(data['sender_name']) == False or Validator.validate_subject(data['subject']) == False or Validator.validate_name(data['sender_email']) == False:
+            for par in req_params:
+                if par not in data.keys():
+                    response = {
+                        "code": 400,
+                        "status": "failure",
+                        "message": "Incomplete Request Body!"
+                    }
+                    return jsonify(response), 400
+            
+            user_id = g.user
+            app_id = data['app_id']
+            name = data['name']
+            subject = data['subject']
+            sender_name = data['sender_name']
+            sender_email = data['sender_email']
+            body = data['body']
+            is_html = data['is_html']
+            
+            check_flag = 1
+            error_str = 'Invalid Data!'
+            error_val = None
+                
+            error_val =  Validator.validate_name(name)
+            
+            if error_val != None:
+                error_str += error_val
+                check_flag = 0
+                
+            error_val =  Validator.validate_name(sender_name)
+            
+            if error_val != None:
+                error_str += error_val
+                check_flag = 0
+                
+            error_val =  Validator.validate_subject(subject)
+            
+            if error_val != None:
+                error_str += error_val
+                check_flag = 0
+                
+            error_val =  Validator.validate_email(sender_email)
+            
+            if error_val != None:
+                error_str += error_val
+                check_flag = 0
+            
+            if check_flag == 0:
                 response = {
                     "code": 400,
                     "status": "failure",
-                    "message": "Invalid Data!"
+                    "message": error_str
                 }
                 return jsonify(response), 400
             
-            if (data['is_html'] == False and HTMLParser.is_valid_text(data['body']) == False) or (data['is_html'] == True and HTMLParser.is_valid_html(data['body']) == False):
+            if (is_html == False and HTMLParser.is_valid_text(body) == False) or (is_html == True and HTMLParser.is_valid_html(body) == False):
                 response = {
                     "code": 400,
                     "status": "failure",
-                    "message": "Invalid Body!"
+                    "message": "Invalid Data! Invalid Body!"
+                }
+                return jsonify(response), 400
+            
+            app_data = App.query.filter_by(app_id=app_id, user_id=user_id).first()
+            
+            if not app_data:
+                response = {
+                    "code": 400,
+                    "status": "failure",
+                    "message": "App does not Exists!"
                 }
                 return jsonify(response), 400
                 
@@ -117,7 +209,7 @@ class TemplatesController:
             response = {
                 "code": 201,
                 "status": "success",
-                "message": "App Created Successfully!",
+                "message": "Template Created Successfully!",
                 "template": new_template.to_dict()
             }
             
@@ -136,17 +228,63 @@ class TemplatesController:
         
         try:
             
-            user_id = g.user
+            req_params = ['template_id', 'name', 'subject', 'sender_name', 'sender_email', 'body', 'is_html']
             
-            if Validator.validate_name(data['name']) == False or Validator.validate_name(data['sender_name']) == False or Validator.validate_subject(data['subject']) == False or Validator.validate_name(data['sender_email']) == False:
+            for par in req_params:
+                if par not in data.keys():
+                    response = {
+                        "code": 400,
+                        "status": "failure",
+                        "message": "Incomplete Request Body!"
+                    }
+                    return jsonify(response), 400
+            
+            user_id = g.user
+            template_id = data['template_id']
+            name = data['name']
+            subject = data['subject']
+            sender_name = data['sender_name']
+            sender_email = data['sender_email']
+            body = data['body']
+            is_html = data['is_html']
+            
+            check_flag = 1
+            error_str = 'Invalid Data!'
+            error_val = None
+                
+            error_val =  Validator.validate_name(name)
+            
+            if error_val != None:
+                error_str += error_val
+                check_flag = 0
+                
+            error_val =  Validator.validate_name(sender_name)
+            
+            if error_val != None:
+                error_str += error_val
+                check_flag = 0
+                
+            error_val =  Validator.validate_subject(subject)
+            
+            if error_val != None:
+                error_str += error_val
+                check_flag = 0
+                
+            error_val =  Validator.validate_email(sender_email)
+            
+            if error_val != None:
+                error_str += error_val
+                check_flag = 0
+            
+            if check_flag == 0:
                 response = {
                     "code": 400,
                     "status": "failure",
-                    "message": "Invalid Data!"
+                    "message": error_str
                 }
                 return jsonify(response), 400
             
-            if (data['is_html'] == False and HTMLParser.is_valid_text(data['body']) == False) or (data['is_html'] == True and HTMLParser.is_valid_html(data['body']) == False):
+            if (is_html == False and HTMLParser.is_valid_text(body) == False) or (is_html == True and HTMLParser.is_valid_html(body) == False):
                 response = {
                     "code": 400,
                     "status": "failure",
@@ -154,7 +292,7 @@ class TemplatesController:
                 }
                 return jsonify(response), 400
             
-            template = Templates.query.filter_by(template_id=data['template_id']).first()
+            template = TemplatesController.validate_template(template_id, user_id)
             
             if not template:
                 response = {
@@ -164,12 +302,12 @@ class TemplatesController:
                 }
                 return jsonify(response), 400
                 
-            template.name = data.get('name', template.name)
-            template.subject = data.get('subject', template.subject)
-            template.sender_name = data.get('sender_name', template.sender_name)
-            template.sender_email = data.get('sender_email', template.sender_email)
-            template.body = data.get('body', template.body)
-            template.is_html = data.get('is_html', template.is_html)
+            template.name = name  if name else template.name
+            template.subject = subject if subject else template.subject
+            template.sender_name = sender_name if sender_name else template.sender_name
+            template.sender_email = sender_email if sender_email else template.sender_email
+            template.body = body if body else template.body
+            template.is_html = is_html if is_html else template.is_html
             db.session.commit()
             
             response = {
@@ -194,9 +332,30 @@ class TemplatesController:
     
         try:
             
-            user_id = g.user
+            req_params = ['template_id', 'password']
             
-            template = Templates.query.filter_by(template_id=data['template_id']).first()
+            for par in req_params:
+                if par not in data.keys():
+                    response = {
+                        "code": 400,
+                        "status": "failure",
+                        "message": "Incomplete Request Body!"
+                    }
+                    return jsonify(response), 400
+            
+            user_id = g.user
+            template_id = data['template_id']
+            password = data['password']
+            
+            if Validator.validate_password(password) != None:
+                response = {
+                    "code": 400,
+                    "status": "failure",
+                    "message": "Invalid Data! Invalid Password!"
+                }
+                return jsonify(response), 400
+            
+            template = TemplatesController.validate_template(template_id, user_id)
             
             if not template:
                 response = {
@@ -206,18 +365,39 @@ class TemplatesController:
                 }
                 return jsonify(response), 400
             
+            existing_user = User.query.filter_by(id=user_id).first()
+            if not existing_user:
+                response = {
+                    "code": 400,
+                    "status": "failure",
+                    "message": "User does not Exists!"
+                }
+                return jsonify(response), 400
+            
+            password_hash = existing_user.get_password_hash()
+            
+            if not Encryptor.verify_password(password_hash, password) :
+                response = {
+                    "code": 400,
+                    "status": "failure",
+                    "message": "Invalid Password!"
+                }
+                return jsonify(response), 400
+            
+            logs = MailLog.query.filter_by(template_id=template_id).all()
+            for log in logs:
+                db.session.delete(log)
+            
             db.session.delete(template)
             db.session.commit()
             
-            response = make_response(
-                    jsonify({
+            response = {
                     "code": 200,
                     "status": "success",
                     "message": "Template Deleted Successfully!"
-                }), 200
-            )
+                }
             
-            return response
+            return jsonify(response), 200
         
         except Exception as e:
             response = {
